move accessor to correct folder
